Add a VPC endpoint for S3
The lack of any VPC endpoints is a non-trivial cost – about $400 a month
in managed NAT Gateway, and that was before anybody started using
Archivematica for serious work.

This change adds an S3 VPC endpoint to the workflow and workflow-stage
VPC, which should dramatically cut those costs.

I haven't set up other endpoints yet because I'm not sure what other
services we're using; I'm guessing S3 is probably the bulk of the load.